refactor(merk): move prove_count_offset_on_range into its own file + version-gate at 0
The method now lives in its own file (`merk/src/merk/prove_count_offset.rs`)
with a `check_merk_v0_with_cost!` gate at the entry. The split keeps
the version contract immediately visible at the file level and
isolates the method from the rest of `Merk::prove*` so future
behavior changes can be reviewed in isolation.

## Versioning

Added a new `MerkProofVersions` struct under `MerkVersions` with a
`prove_count_offset_on_range: FeatureVersion` field. Initial
implementation version is **0**, set across all shipped grove
versions (v1, v2, v3):

- **v3**: 0 — the method is reachable here via the V1 proof envelope
  (which honors count-offset paginated queries).
- **v1 / v2**: 0 — the method is *not* reachable from these versions
  in normal use because their V0 proof envelope rejects offsets at
  the grovedb dispatch layer before reaching merk. The field is
  kept consistent so a direct merk caller (outside the grovedb
  proof dispatch) doesn't trip the version gate spuriously.

Bumping `prove_count_offset_on_range` from 0 → 1 in a future grove
version is the prescribed path if the prover's emitted op stream
needs to change shape in a way that requires a coordinated verifier
update.

## Files

- New: `merk/src/merk/prove_count_offset.rs`
- New struct: `MerkProofVersions` in `grovedb-version/src/version/merk_versions.rs`
- Updated: `merk/src/merk/mod.rs` registers the new module
- Updated: `merk/src/merk/prove.rs` no longer holds the method (a
  pointer comment in its place)
- Updated: `grovedb-version/src/version/v{1,2,3}.rs` carry the new
  field

No behavior change for any caller — the method's body is byte-identical
to the previous version, just gated and relocated.

Test totals: merk count_offset 36/36, grovedb count_offset 24/24,
full merk 529/529, full grovedb 1737/1737.
