+++ merk/src/merk/prove.rs
@@ -5,11 +5,7 @@ use grovedb_storage::StorageContext;
 use grovedb_version::version::GroveVersion;
 
 use crate::{
-    proofs::{
-        encode_into,
-        query::{count_offset::ProverCountOffsetResult, QueryItem},
-        Op as ProofOp, Query,
-    },
+    proofs::{encode_into, query::QueryItem, Op as ProofOp, Query},
     tree::RefWalker,
     Error, Merk,
 };
@@ -189,74 +185,10 @@ where
         })
     }
 
-    /// Generate an offset-paginated proof for a single-range query
-    /// against a `ProvableCountTree` or `ProvableCountSumTree`.
-    ///
-    /// This is the count-tree analogue of the regular [`Self::prove`]
-    /// path, with one key extension: a non-zero `offset` is honored.
-    /// The proof commits the count of skipped items via the same
-    /// `HashWithCount` infrastructure used by
-    /// [`Self::prove_aggregate_count_on_range`], so the offset region
-    /// pays O(log n) proof size per skipped subtree rather than
-    /// O(skipped). Returned items inside the limit window emit as
-    /// normal count-bearing value nodes, so the verifier-side result
-    /// shape matches what a regular range query without offset would
-    /// produce.
-    ///
-    /// `inner_range` is the single `QueryItem` to scan (already
-    /// validated at the caller's `Query`/`PathQuery` level). `offset`
-    /// is how many leading in-range items to skip (in directional
-    /// order); `limit` is the maximum number of items to return after
-    /// the offset (`None` means unlimited). `left_to_right` controls
-    /// iteration direction.
-    ///
-    /// The merk's `tree_type` must be one of `ProvableCountTree` /
-    /// `ProvableCountSumTree`. Any other tree type is rejected with
-    /// `Error::InvalidProofError` before any walking happens — count
-    /// commitments are only meaningful against trees that bind their
-    /// count into the node hash. Empty merk: returns an empty
-    /// `ProverCountOffsetResult` (no ops, 0 returned, full offset
-    /// remaining).
-    pub fn prove_count_offset_on_range(
-        &self,
-        inner_range: &QueryItem,
-        offset: u64,
-        limit: Option<u64>,
-        left_to_right: bool,
-        grove_version: &GroveVersion,
-    ) -> CostResult<ProverCountOffsetResult, Error> {
-        let tree_type = self.tree_type;
-        if !matches!(
-            tree_type,
-            crate::TreeType::ProvableCountTree | crate::TreeType::ProvableCountSumTree
-        ) {
-            return Err(Error::InvalidProofError(format!(
-                "count-offset paginated proof is only valid against ProvableCountTree or \
-                 ProvableCountSumTree, got {:?}",
-                tree_type
-            )))
-            .wrap_with_cost(Default::default());
-        }
-        self.use_tree_mut(|maybe_tree| match maybe_tree {
-            None => Ok(ProverCountOffsetResult {
-                ops: LinkedList::new(),
-                returned: 0,
-                offset_remaining: offset,
-            })
-            .wrap_with_cost(Default::default()),
-            Some(tree) => {
-                let mut ref_walker = RefWalker::new(tree, self.source());
-                ref_walker.create_count_offset_on_range_proof(
-                    inner_range,
-                    offset,
-                    limit,
-                    left_to_right,
-                    tree_type,
-                    grove_version,
-                )
-            }
-        })
-    }
+    // `prove_count_offset_on_range` lives in
+    // `merk/src/merk/prove_count_offset.rs` — it's a version-gated
+    // entry point (`MerkProofVersions::prove_count_offset_on_range`)
+    // and the split keeps that contract immediately visible.
 
     /// Generate a sum-only proof for an `AggregateSumOnRange` query.
     /// Mirror of [`Self::prove_aggregate_count_on_range`] for the
